Extracted the nodes thing to its own package
